% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alignment_successful}
\alias{alignment_successful}
\title{A function that indicates if an alignment was successful}
\usage{
alignment_successful(epitope, alignment)
}
\arguments{
\item{epitope}{The epitope that was searched for in the reference}

\item{alignment}{The alignment to check - usually the output from
pairwiseAlignment}
}
\description{
A function that indicates if an alignment was successful
}

