% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{BETWEEN10}
\alias{BETWEEN10}
\title{HIV Scoring Matrix 10 for between host evolution}
\format{
A 24x24 matrix
}
\usage{
data(BETWEEN10)
}
\description{
HIV Scoring Matrix 10 for between host evolution
}
\keyword{datasets}
