% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{BETWEEN38}
\alias{BETWEEN38}
\title{HIV Scoring Matrix 38 for between host evolution}
\format{
A 24x24 matrix
}
\usage{
data(BETWEEN38)
}
\description{
HIV Scoring Matrix 38 for between host evolution
}
\keyword{datasets}
