% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_scoring_jobs}
\alias{build_scoring_jobs}
\title{Builds scoring jobs from pre-processed inputs}
\usage{
build_scoring_jobs(matched_patients, flat_lanl_hla)
}
\arguments{
\item{matched_patients}{As produced by
\code{\link{match_patient_hla_to_query_alignment}}.}

\item{flat_lanl_hla}{As produced by \code{\link{flatten_lanl_hla}}}
}
\description{
Called by \code{\link{list_scores_to_compute }}
}

