% Generated by roxygen2 (4.0.2): do not edit by hand
\name{epitope_found}
\alias{epitope_found}
\title{A function that indicates if an epitope was found in the reference sequence.}
\usage{
epitope_found(epitope, alignment)
}
\arguments{
\item{epitope}{The epitope that was searched for in the reference}

\item{alignment}{The alignment to check - usually the output from
pairwiseAlignment}
}
\description{
An epitope is seen as not found if the alignment between the epitope and the
reference does not contain the same number of bases as the epitope. This
should maybe be relaxed in the future.
}

