% Generated by roxygen2 (4.0.2): do not edit by hand
\name{epitope_pos_in_ref}
\alias{epitope_pos_in_ref}
\title{Finds the position of the epitope in the reference sequence}
\usage{
epitope_pos_in_ref(epitope, query_alignment)
}
\arguments{
\item{epitope}{The epitope to find in the sequence. Either a character
string or an AAString}

\item{query_alignment}{The query alignment}
}
\description{
It uses pairwiseAlignment with the default settings. See Biostrings manual
}
\details{
Lots of things to test and investigate that can improve this function.
}

