% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list_jobs.R
\name{flatten_lanl_hla}
\alias{flatten_lanl_hla}
\title{Flattens the LANL HLA file}
\usage{
flatten_lanl_hla(lanl_hla)
}
\arguments{
\item{lanl_hla}{The data.frame (of class LANL_HLA_data) that contains
the descriptions of the different HLA genotypes}
}
\description{
Sometimes the same hla has a number of different names. Hence the
hla_genotype column in the lanl file must be processed before the matches
can be made.
}
\details{
This function takes the LANL HLA file and transforms it so that each row
correspond to one and only one hla. This means that row in which the
'hla_genotype' column is unpopulated gets discarded and that rows in which
the 'hla_genotype' column contains the names of more than one hla_genotype
(assumed to be seperated by commas) will be duplicated and each duplicate
will be assigned to one hla_genotype.
}

