% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get_epitope}
\alias{get_epitope}
\alias{get_epitope,Scoring_Job-method}
\title{Returns the epitope associated with a scoring_job}
\usage{
get_epitope(the_scoring_job)

\S4method{get_epitope}{Scoring_Job}(the_scoring_job)
}
\arguments{
\item{the_scoring_job}{The scoring job which epitope must be extracted}
}
\description{
Returns the epitope associated with a scoring_job
}

