% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get_hla_details}
\alias{get_hla_details}
\alias{get_hla_details,Scoring_Job-method}
\title{Returns the hla details associated with a scoring_job}
\usage{
get_hla_details(the_scoring_job)

\S4method{get_hla_details}{Scoring_Job}(the_scoring_job)
}
\arguments{
\item{the_scoring_job}{The scoring job whose details must be extracted}
}
\description{
Returns the hla details associated with a scoring_job
}

