% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scoring_job.R
\docType{methods}
\name{get_hla_genotype}
\alias{get_hla_genotype}
\alias{get_hla_genotype,Scoring_Job-method}
\title{Returns the hla genotype associated with a scoring_job}
\usage{
get_hla_genotype(the_scoring_job)

\S4method{get_hla_genotype}{Scoring_Job}(the_scoring_job)
}
\arguments{
\item{the_scoring_job}{The scoring job which epitope must be extracted}
}
\description{
Returns the hla genotype associated with a scoring_job
}

