% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get_matchable_patient_ids}
\alias{get_matchable_patient_ids}
\alias{get_matchable_patient_ids,AAStringSet,Patient_HLA-method}
\title{Returns the patient_ids that are in both the query_alignment and the
patient_hla file.}
\usage{
get_matchable_patient_ids(query_alignment, patient_hla)

\S4method{get_matchable_patient_ids}{AAStringSet,Patient_HLA}(query_alignment,
  patient_hla)
}
\arguments{
\item{query_alignment}{The query alignment}

\item{patient_hla}{The data.frame (of class Patient_HLA) that contain lists
all the HLA genotypes each patient has.}
}
\description{
Returns the patient_ids that are in both the query_alignment and the
patient_hla file.
}

