% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_matchable_patient_ids}
\alias{get_matchable_patient_ids}
\title{Gets all the patient ids that are in both the MSA and the patient_hla file}
\usage{
get_matchable_patient_ids(query_alignment, patient_hla)
}
\arguments{
\item{query_alignment}{The query alignment}

\item{patient_hla}{The data.frame (of class Patient_HLA) that contain lists
all the HLA genotypes each patient has.}
}
\description{
Gets all the patient ids that are in both the MSA and the patient_hla file
}

