% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_set_of_test_data}
\alias{get_set_of_test_data}
\title{Get full set of test data}
\usage{
get_set_of_test_data(output_folder = ".")
}
\arguments{
\item{output_folder}{Folder in which the files are to be produced. Defaults to the current folder.}
}
\description{
Running this function will produce 3 files in the specified folder.
\enumerate{
  \item test_patient_hla_genotypes.csv as produced by get_test_patient_hla_data()
  \item test_lanl_hla_data.csv as produced by get_test_lanl_hla_data()
  \item test_query_alignment.fasta as produced by get_test_query_alignment()
}
}
