% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_test_lanl_hla_data}
\alias{get_test_lanl_hla_data}
\title{A function that returns a test lanl hla genotype dataset}
\usage{
get_test_lanl_hla_data(dataset_name = "base")
}
\arguments{
\item{dataset_name}{The name of the test dataset to return}
}
\description{
A function that returns a test lanl hla genotype dataset
}

