% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_HLA.R
\name{get_test_patient_hla_data}
\alias{get_test_patient_hla_data}
\title{A function that returns a test patient hla dataset}
\usage{
get_test_patient_hla_data(dataset_name = "base")
}
\arguments{
\item{dataset_name}{The name of the test dataset to return}
}
\description{
A function that returns a test patient hla dataset
}
