% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{get_test_query_alignment}
\alias{get_test_query_alignment}
\title{A function that returns a test query alignment}
\usage{
get_test_query_alignment(dataset_name = "base")
}
\arguments{
\item{dataset_name}{The name of the test dataset to return}
}
\description{
A function that returns a test query alignment
}
