% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{LANL_HLA_data-class}
\alias{.LANL_HLA_data}
\alias{LANL_HLA_data-class}
\title{The class for the data from LANL that describe the HLA genotype's epitopes}
\description{
The class for the data from LANL that describe the HLA genotype's epitopes
}

