% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{list_epitopes}
\alias{list_epitopes}
\alias{list_epitopes,AAStringSet,Patient_HLA,LANL_HLA_data-method}
\title{Lists all the hla genotypes that must be investigated}
\usage{
list_epitopes(query_alignment, patient_hla, lanl_hla_data)


  \S4method{list_epitopes}{AAStringSet,Patient_HLA,LANL_HLA_data}(query_alignment,
  patient_hla, lanl_hla_data)
}
\arguments{
\item{query_alignment}{The query alignment}

\item{patient_hla}{The data.frame (of class Patient_HLA) that contain lists
all the HLA genotypes each patient has.}

\item{lanl_hla_data}{The data.frame (of class LANL_HLA_data) that contains
the descriptions of the different HLA genotypes}
}
\description{
Lists all the hla genotypes that must be investigated
}

