% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_epitopes}
\alias{list_epitopes}
\title{Lists all the hla genotypes that must be investigated}
\usage{
list_epitopes(query_alignment, patient_hla, lanl_hla_data)
}
\arguments{
\item{query_alignment}{The query alignment}

\item{patient_hla}{The data.frame (of class Patient_HLA) that contain lists
all the HLA genotypes each patient has.}
}
\description{
Lists all the hla genotypes that must be investigated
}

