% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{list_hlas}
\alias{list_hlas}
\alias{list_hlas,AAStringSet,Patient_HLA-method}
\title{Lists all the hla genotypes that must be investigated}
\usage{
list_hlas(query_alignment, patient_hla)

\S4method{list_hlas}{AAStringSet,Patient_HLA}(query_alignment, patient_hla)
}
\arguments{
\item{query_alignment}{The query alignment}

\item{patient_hla}{The data.frame (of class Patient_HLA) that contain lists
all the HLA genotypes each patient has.}
}
\description{
Lists all the hla genotypes that must be investigated
}

