% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_jobs.R
\name{list_scores_to_compute}
\alias{list_scores_to_compute}
\title{Processes the three input files (query_alignment, patient_hla and lanl_hla)
in to a list of scoring jobs.}
\usage{
list_scores_to_compute(query_alignment, patient_hla, lanl_hla)
}
\arguments{
\item{query_alignment}{An AAStringSet that contains the multiple sequence
alignment of the patient's viral sequences}

\item{patient_hla}{The data.frame that specifies which query sequence to
check against which hla genotypes}

\item{lanl_hla}{The data.frame (of class LANL_HLA_data) that contains
the descriptions of the different HLA genotypes}
}
\value{
A list of Scoring_Jobs
}
\description{
First the patient_hla data is matched to the query_alignment
\code{\link{match_patient_hla_to_query_alignment}}, then the lanl_hla file
is flattened \code{\link{flatten_lanl_hla}}, and finally, the jobs are built
\code{\link{build_scoring_jobs}}.
}
\details{
This list of jobs can then be used to perform the comparisons.
}
