% Generated by roxygen2 (4.0.2): do not edit by hand
\name{match_epitopes}
\alias{match_epitopes}
\title{Computes similarities between certain epitopes and sequences}
\usage{
match_epitopes(query_alignment, patient_hla, lanl_hla_data,
  range_expansion = 5, update_progress_bar = NULL,
  substitutionMatrix = "BLOSUM50")
}
\arguments{
\item{query_alignment}{The query alignment}

\item{patient_hla}{The data.frame (of class Patient_HLA) that contain lists
all the HLA genotypes each patient has.}

\item{lanl_hla_data}{The data.frame (of class LANL_HLA_data) that contains
the descriptions of the different HLA genotypes}

\item{range_expansion}{After the epitope is found in the reference
seqeuence, search in each of the query sequences for the same epitope, but
expand the range with this number of amino acids}

\item{update_progress_bar}{A closure passed in from a reactive shiny
expression that allows a progress bar to be updated when using the shiny web
ui.}

\item{substitutionMatrix}{substitution matrix representing the fixed substitution
scores for an alignment. It cannot be used in conjunction with ‘patternQuality’
and ‘subjectQuality’ arguments.}
}
\description{
A query alignment and a file specifying which hla_genotypes should be
checked for different patients are first compared to construct a list of
scores that must be computed. This list is then passed to the
score_all_epitopes function which computes the scores. The results and error
logs are then returned as output.
}

