% Generated by roxygen2 (4.0.2): do not edit by hand
\name{match_patient_hla_to_lanl_hla}
\alias{match_patient_hla_to_lanl_hla}
\title{Extracts an hla's details from the LANL file}
\usage{
match_patient_hla_to_lanl_hla(patient_hla, lanl_hla)
}
\arguments{
\item{patient_hla}{The data.frame that specifies which query sequence to
check against which hla genotypes}

\item{lanl_hla}{The data.frame (of class LANL_HLA_data) that contains
the descriptions of the different HLA genotypes}
}
\value{
A list of data.frames of hla details. The list is indexed by
hla_genotype and the data.frame contains a single row with the relevant
details of the hla_genotype.
}
\description{
Extracts an hla's details from the LANL file
}

