% Generated by roxygen2 (4.0.2): do not edit by hand
\name{match_patient_hla_to_query_alignment}
\alias{match_patient_hla_to_query_alignment}
\title{Matches the patient hla data to the query sequence to check which hla's must
be checked for in which sequences.}
\usage{
match_patient_hla_to_query_alignment()
}
\description{
It treats the patient_id column in the patient_hla data as a regular
expression and performs the lookup to the query sequence names
}

