% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Patient_HLA-class}
\alias{.Patient_HLA}
\alias{Patient_HLA-class}
\title{The class for the data that describes which patients have which HLAs}
\description{
The class for the data that describes which patients have which HLAs
}

