% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_lanl_hla}
\alias{read_lanl_hla}
\title{A function that reads a HLA genotype specification file}
\usage{
read_lanl_hla(file_name)
}
\arguments{
\item{file_name}{Name of the file}
}
\description{
This file is typically obtained from
http://www.hiv.lanl.gov/content/immunology/tables/tables.html.
}
\details{
The file must have the following columns:
\itemize{
 \item{Epitope}
 \item{Protein}
 \item{HXB2.start}
 \item{HXB2.end}
 \item{Subprotein}
 \item{HXB2.DNA.Contig}
 \item{Subtype}
 \item{Species}
 \item{HLA}
}

The input format should be made more flexible in the future.
}

