% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_patient_hla}
\alias{read_patient_hla}
\title{A function that reads a patient HLA genotype specification file}
\usage{
read_patient_hla(file_name)
}
\arguments{
\item{file_name}{Name of the file}
}
\description{
The file must have two columns:
\itemize{
 \item{patient_id}
 \item{hla_genotype}
}
}
\details{
The value of the patient_id column gets treated as a regular expression
and is matched to the FASTA headers in the query_alignment. If the value is
set to .* then the hla_genotype corresponding to this entry will be matched
to all the sequences in the query alignment.

The hla_genotype column should be a name from the LANL table.
}

