% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_query_alignment}
\alias{read_query_alignment}
\title{Reads in the query alignment}
\usage{
read_query_alignment(file_name)
}
\arguments{
\item{file_name}{Name of the fasta file}
}
\description{
This function currently just calls readAAStringSet in the Biostrings
package. See that function for more details.
}
\details{
Must be a valid FASTA file.
}

