% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{run_EpitopeMatcher_app}
\alias{run_EpitopeMatcher_app}
\title{Runs the shiny app for this package}
\usage{
run_EpitopeMatcher_app(port = 5436)
}
\description{
Runs the shiny app for this package
}

