% Generated by roxygen2 (4.0.2): do not edit by hand
\name{score_all_epitopes}
\alias{score_all_epitopes}
\title{Scores how well the epitopes in a patient's virus' sequences will be recognized by
the patient's HLA genotype.}
\usage{
score_all_epitopes(query_alignment, patient_hla, lanl_hla_data,
  range_expansion = 5)
}
\arguments{
\item{query_alignment}{The query alignment}

\item{patient_hla}{The data.frame (of class Patient_HLA) that contain lists
all the HLA genotypes each patient has.}

\item{lanl_hla_data}{The data.frame (of class LANL_HLA_data) that contains
the descriptions of the different HLA genotypes}

\item{range_expansion}{After the epitope is found in the reference
seqeuence, search in each of the query sequences for the same epitope, but
expand the range with this number of amino acids}
}
\description{
Scores how well the epitopes in a patient's virus' sequences will be recognized by
the patient's HLA genotype.
}

