% Generated by roxygen2 (4.0.2): do not edit by hand
\name{score_sequence_epitopes}
\alias{score_sequence_epitopes}
\title{Scores how well the epitopes in a patient's virus' sequences will be recognized by
the patient's HLA genotype.}
\usage{
score_sequence_epitopes(query_alignment, patient_hla, lanl_hla_data,
  range_expansion = 5)
}
\arguments{
\item{query_alignment}{The query alignment}

\item{patient_hla}{The data.frame (of class Patient_HLA) that contain lists
all the HLA genotypes each patient has.}

\item{lanl_hla_data}{The data.frame (of class LANL_HLA_data) that contains
the descriptions of the different HLA genotypes}
}
\description{
Call for testing:
score_sequence_epitopes(read_query_alignment(), read_patient_hla(),
read_lanl_hla())
}

