% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Scoring_Job-class}
\alias{.Scoring_Job}
\alias{Scoring_Job-class}
\title{A class that contains all the required information to run an epitope scoring
job.}
\description{
Three pieces of information is required for a scoring job to be valid:
\begin{itemize}
  \item{hla_genotype - The name of the hla_genotype to investigate}
  \item{query_sequence_names - A character vector of the names of the query
  sequences in which this hla_genotype must be looked for}
  \item{hla_details - A list of further details about this hla_genotype}
\end{itemize}
}

