% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{WITHIN5}
\alias{WITHIN5}
\title{HIV Scoring Matrix 5 for within host evolution}
\format{A 24x24 matrix}
\usage{
data(WITHIN5)
}
\description{
HIV Scoring Matrix 5 for within host evolution
}
\keyword{datasets}

