% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute_epitope_scores}
\alias{compute_epitope_scores}
\title{Computes the similarities between the epitope and the sequences in the
alignment}
\usage{
compute_epitope_scores(epitope, query_alignment)
}
\arguments{
\item{epitope}{The epitope to find in the sequence. Either a character
string or an AAString}

\item{query_alignment}{The query alignment}
}
\description{
Computes the similarities between the epitope and the sequences in the
alignment
}

