% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_HLA.R, R/msa.R
\name{get_patient_ids}
\alias{get_patient_ids}
\alias{get_patient_ids,Patient_HLA-method}
\alias{get_patient_ids,AAStringSet-method}
\title{Returns the ids of the patients in the data structure}
\usage{
get_patient_ids(x, sep = "\\\\|", id_position = 1)

\S4method{get_patient_ids}{Patient_HLA}(x)

\S4method{get_patient_ids}{AAStringSet}(x, sep = "\\\\|", id_position = 1)
}
\arguments{
\item{x}{The data structure to interrogate}

\item{sep}{The symbol used to separate elements in the sequence names}

\item{id_position}{After the sequence name has been split on the 'sep'
character, which element of the resulting vector contains the patient id?}
}
\description{
It parses the FASTA header when called on a XStringSet object using the sep
and id_position arguments.
}
