% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get_patient_ids,AAMultipleAlignment-method}
\alias{get_patient_ids}
\alias{get_patient_ids,AAMultipleAlignment-method}
\alias{get_patient_ids,Patient_HLA-method}
\title{Returns the ids of the patients in the data structure}
\usage{
\S4method{get_patient_ids}{AAMultipleAlignment}(x)

get_patient_ids(x)

\S4method{get_patient_ids}{Patient_HLA}(x)
}
\arguments{
\item{data}{The data structure to interrogate}
}
\description{
Returns the ids of the patients in the data structure
}

