% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{LANL_HLA_data-class}
\alias{.LANL_HLA_data}
\alias{LANL_HLA_data-class}
\title{The class for the data from LANL that describe the HLA genotype's epitopes}
\description{
This class is an extension of the data.frame class placing some extra
restrictions on the data format. This data is typically obtained from
http://www.hiv.lanl.gov/content/immunology/tables/tables.html.
}
\details{
The data.frame has the following columns:
\itemize{
 \item{Epitope}
 \item{Protein}
 \item{HXB2.start}
 \item{HXB2.end}
 \item{Subprotein}
 \item{HXB2.DNA.Contig}
 \item{Subtype}
 \item{Species}
 \item{HLA}
}
}

