% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_jobs.R
\name{match_patient_hla_to_query_alignment}
\alias{match_patient_hla_to_query_alignment}
\title{Matches the patient_hla data to the query sequence to check which hla's must
be checked for in which sequences.}
\usage{
match_patient_hla_to_query_alignment(query_alignment, patient_hla)
}
\arguments{
\item{query_alignment}{An AAStringSet that contains the multiple sequence
alignment of the patient's viral sequences}

\item{patient_hla}{The data.frame that specifies which query sequence to
check against which hla genotypes}
}
\value{
A list of lists. The inner lists contains the elements
'hla_genotype' and 'query_sequence_names'
}
\description{
It treats the patient_id column in the patient_hla data as a regular
expression and performs the lookup to the query sequence names.
}
