% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_lanl_hla}
\alias{read_lanl_hla}
\title{A function that reads a patient HLA genotype specification file}
\usage{
read_lanl_hla(file_name)
}
\arguments{
\item{file_name}{Name of the file}
}
\description{
A function that reads a patient HLA genotype specification file
}

