% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_lanl_hla}
\alias{read_lanl_hla}
\title{A function that reads a HLA genotype specification file}
\usage{
read_lanl_hla(file_name)
}
\arguments{
\item{file_name}{Name of the file}
}
\description{
This function converts the file into an object of class LANL_HLA. See the
documentation of that function for more details:
\code{\link{.LANL_HLA_data}}
}

