% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/patient_HLA.R
\name{read_patient_hla}
\alias{read_patient_hla}
\title{A function that reads a patient HLA genotype specification file.}
\usage{
read_patient_hla(file_name)
}
\arguments{
\item{file_name}{Name of the file}
}
\description{
The function will convert the file into an object of class Patient_HLA.
See the class documentation for more details about the format:
\code{\link{.Patient_HLA}}
}

