% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{read_query_alignment}
\alias{read_query_alignment}
\title{Reads in the query alignment}
\usage{
read_query_alignment(file_name)
}
\arguments{
\item{file_name}{Name of the fasta file}
}
\description{
This function currently just calls readAAStringSet in the Biostrings
package. See that function for more details.
}
\details{
Must be a valid FASTA file.

The FASTA headers must be in some delimited form with a special character
used for delimiting different fields. Further more, the patient id must
always be in the same column in the FASTA header. For example:
PATID_TIME_PID_CONSENSUSDETAILS
}
