% Generated by roxygen2 (4.0.2): do not edit by hand
\name{score_all_epitopes}
\alias{score_all_epitopes}
\title{Given a list of scoring jobs, compute the similarities}
\usage{
score_all_epitopes(the_scoring_jobs, query_alignment, range_expansion)
}
\arguments{
\item{the_scoring_jobs}{A list of scoring_jobs}

\item{query_alignment}{The query alignment}

\item{range_expansion}{After the epitope is found in the reference
seqeuence, search in each of the query sequences for the same epitope, but
expand the range with this number of amino acids}
}
\description{
Given a list of scoring jobs, compute the similarities
}

