% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_job.R
\docType{class}
\name{Scoring_Job-class}
\alias{Scoring_Job-class}
\alias{.Scoring_Job}
\title{A class that contains all the required information to run an epitope scoring
job.}
\description{
Three pieces of information is required for a scoring job to be valid:
\itemize{
  \item{hla_genotype - The name of the hla_genotype to investigate}
  \item{query_sequence_names - A character vector of the names of the query
  sequences in which this hla_genotype must be looked for}
  \item{hla_details - A list of further details about this hla_genotype}
}
}
\details{
The hla_details is forced to have the following values: "end_pos", 
"epitope", "gene_name", "hla_genotype", "hxb2_dna_position", "organism", 
"start_pos", "subprotein", and "sub_type". This is reasonable since the HLA
data will always come from the LANL file and these details must be added to
the results and error logs.
}
